rewrite ContractView using MVVM

+++ AccountsReceivable/Data/Repositories/CompanyRepository.cs
@@ -1,17 +1,16 @@
 ﻿using AccountsReceivable.Data.Interfaces;
+using AccountsReceivable.Exceptions;
 using AccountsReceivable.Models;
 using AccountsReceivable.Services;
+using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
-using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace AccountsReceivable.Data.Repositories
 {
-    public class CompanyRepository : ICompanyRepository
+    public class CompanyRepository : NotifiableRepository, ICompanyRepository
     {
         private readonly IDbContextFactory<ApplicationContext> factory;
         public CompanyRepository(IDbContextFactory<ApplicationContext> factory) 
@@ -23,11 +22,12 @@ namespace AccountsReceivable.Data.Repositories
             using var context = await factory.CreateDbContextAsync();
             return await context.Companies.FindAsync(id);
         }
-        public async Task<int> AddCompanyAsync(Company company)
+        public async Task<int> CreateCompanyAsync(Company company)
         {
             using var context = await factory.CreateDbContextAsync();
             await context.Companies.AddAsync(company);
             await context.SaveChangesAsync();
+            await NotifyDataChangedAsync();
             return company.Id;
         }
         public async Task UpdateCompanyAsync(Company company)
@@ -35,18 +35,33 @@ namespace AccountsReceivable.Data.Repositories
             using var context = await factory.CreateDbContextAsync();
             context.Update(company);
             await context.SaveChangesAsync();
+            await NotifyDataChangedAsync();
         }
-        public async Task<IEnumerable<Company>> GetAllCounterpartiesAsync()
+        public async Task<IEnumerable<Company>> GetAllCompaniesAsync()
         {
             using var context = await factory.CreateDbContextAsync();
             return await context.Companies.Where(x => x.Id != Constants.OWN_COMPANY_ID).Include(c => c.Category).ToListAsync();           
         }
-        public async Task RemoveCounterpartiesAsync(List<Company> companies)
+        public async Task DeleteCompaniesAsync(List<Company> companies)
         {
-            using var context = await factory.CreateDbContextAsync();
-            context.Companies.AttachRange(companies);
-            context.Companies.RemoveRange(companies);
-            await context.SaveChangesAsync();
+            if (companies == null || companies.Count == 0) return;
+            try
+            {
+                using var context = await factory.CreateDbContextAsync();
+                context.Companies.AttachRange(companies);
+                context.Companies.RemoveRange(companies);
+                await context.SaveChangesAsync();
+                await NotifyDataChangedAsync();
+            }
+            catch (DbUpdateException ex)
+            {
+                if (ex.InnerException is SqliteException sqliteEx && sqliteEx.SqliteErrorCode == 19)
+                {
+                    var message = companies.Count == 1  ? "Невозможно удалить компанию. Есть связанные записи." : "Невозможно удалить компании. Одна или несколько имеют связанные записи.";
+                    throw new DeleteRestrictedException(message);
+                }
+                throw;
+            }
         }
     }
 }
